load-test/docker-image/locust-tasks/requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
idna==2.6
itsdangerous==0.24
Jinja2==2.10
locustio==0.8.1
MarkupSafe==1.0
msgpack==0.5.0
msgpack-python==0.5.0
pyzmq==16.0.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
ws4py==0.4.3
